# POSTCONDITION is evaluated twice with different values for TLCExt!Counterexample when TLC detects a liveness violation.