@lemmy 2024-10-18
Also, TLC will no longer print `Model checking completed. No error has been found.` if the postcondition is violated:

### Old
```shell
...
Starting... (2024-10-17 19:20:51)
Implied-temporal checking--satisfiability problem has 1 branches.
Computing initial states...
Finished computing initial states: 1 distinct state generated at 2024-10-17 19:20:51.
Error: Assumption line 81, col 5 to line 81, col 54 of module MCCRDT is false.
Progress(1) at 2024-10-17 19:20:51: 8 states generated, 1 distinct states found, 0 states left on queue.
Checking temporal properties for the complete state space with 1 total distinct states at (2024-10-17 19:20:51)
Finished checking temporal properties in 00s at 2024-10-17 19:20:51

Model checking completed. No error has been found.
  Estimates of the probability that TLC did not check all reachable states
  because two distinct states had the same fingerprint:
  calculated (optimistic):  val = 3.8E-19

8 states generated, 1 distinct states found, 0 states left on queue.
The depth of the complete state graph search is 1.
The average outdegree of the complete state graph is 0 (minimum is 0, the maximum 0 and the 95th percentile is 0).
Finished in 00s at (2024-10-17 19:20:51)
```

### New
```shell
...
Starting... (2024-10-17 19:20:54)
Implied-temporal checking--satisfiability problem has 1 branches.
Computing initial states...
Finished computing initial states: 1 distinct state generated at 2024-10-17 19:20:55.
Progress(1) at 2024-10-17 19:20:55: 8 states generated, 1 distinct states found, 0 states left on queue.
Checking temporal properties for the complete state space with 1 total distinct states at (2024-10-17 19:20:55)
Finished checking temporal properties in 00s at 2024-10-17 19:20:55
Error: Assumption line 81, col 5 to line 81, col 54 of module MCCRDT is false.
8 states generated, 1 distinct states found, 0 states left on queue.
The depth of the complete state graph search is 1.
Finished in 00s at (2024-10-17 19:20:55)
```